# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MaLoveApp Backend API
  version: 0.1.0
  description: |
    Especificación OpenAPI 3 del backend de MaLoveApp.

    Seguridad principal: Firebase Auth (Bearer token). En desarrollo puede estar activo el bypass de tokens mock (`ALLOW_MOCK_TOKENS`).

    Documentación relacionada:
    - Arquitectura: ../ARCHITECTURE.md
    - Onboarding: ../ONBOARDING.md
    - Monitoreo: ../monitoring/README.md
servers:
  - url: http://localhost:4004
    description: Local (desarrollo)
  - url: https://maloveapp-backend.onrender.com
    description: Render (staging/producción)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cronToken:
      type: apiKey
      in: header
      name: X-Cron-Token
...
paths:
...
  /api/emails:
    get:
      tags: [emails]
      summary: Lista emails (por carpeta)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: folder
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/emails/{id}:
    get:
      tags: [emails]
      summary: Detalle email
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '404': { description: Not found }
...